Atualização

+++ Source/Posto.Win.Update/Infraestrutura/Atualizar.cs
@@ -33,6 +33,8 @@ namespace Posto.Win.Update.Infraestrutura
 
         private const string PathLeitor = @"C:\Metodos\uLeitor.exe";
 
+        private const string PathPostoWeb = @"C:\Metodos\uPostoWe.exe";
+
         #endregion
 
         #region Propriedades
@@ -49,30 +51,30 @@ namespace Posto.Win.Update.Infraestrutura
 
         public Atualizar(AtualizarModel atualizar)
         {
-            this._atualizar = atualizar;
-            this._ftp = new Ftp();
+            _atualizar = atualizar;
+            _ftp = new Ftp();
 
-            this._timerProximaAtualizacao = new Timer();
-            this._timerProximaAtualizacao.AutoReset = false;
-            this._timerProximaAtualizacao.Elapsed += this.Execute;
+            _timerProximaAtualizacao = new Timer();
+            _timerProximaAtualizacao.AutoReset = false;
+            _timerProximaAtualizacao.Elapsed += Execute;
         }
 
         public void Iniciar(ConfiguracaoModel configuracaoModel)
         {
-            this._configuracaoModel = configuracaoModel;
-            this.SetTime();
+            _configuracaoModel = configuracaoModel;
+            SetTime();
         }
 
         public void Pausar()
         {
-            this._timerProximaAtualizacao.Stop();
+            _timerProximaAtualizacao.Stop();
         }
 
         public async void Manual(ConfiguracaoModel configuracaoModel, DelegateCommand atualizarAfter)
         {
-            this._configuracaoModel = configuracaoModel;
-            await this.ExecuteAsync();
-            this._timerProximaAtualizacao.Stop();
+            _configuracaoModel = configuracaoModel;
+            await ExecuteAsync();
+            _timerProximaAtualizacao.Stop();
             atualizarAfter.Execute();
         }
 
@@ -81,22 +83,30 @@ namespace Posto.Win.Update.Infraestrutura
         /// </summary>
         public void AtualizaViewModel()
         {
-            var context = new PostoContext(this._configuracaoModel);
-            var query = context.Query("SELECT *, (SELECT oft000.parversao FROM oft000 LIMIT 1) AS versao, NOW() AS dataAtual FROM atualiz");
-
-            using (var reader = query.ExecuteReader())
+            try
             {
-                while (reader.Read())
+                var context = new PostoContext(_configuracaoModel);
+                var query = context.Query("SELECT *, (SELECT oft000.parversao FROM oft000 LIMIT 1) AS versao, NOW() AS dataAtual FROM atualiz");
+
+                using (var reader = query.ExecuteReader())
                 {
-                    this._atualizar.Dia         = reader.GetInt16(reader.GetOrdinal("atuDiaSema"));
-                    this._atualizar.Hora        = Convert.ToInt16(reader.GetString(reader.GetOrdinal("atuHoraMin")).Split(':')[0]);
-                    this._atualizar.Minuto      = Convert.ToInt16(reader.GetString(reader.GetOrdinal("atuHoraMin")).Split(':')[1]);
-                    this._atualizar.UltimaData  = reader.GetDateTime(reader.GetOrdinal("atuDatAtua"));
-                    this._atualizar.Versao      = reader.GetInt32(reader.GetOrdinal("versao"));
-                    this._atualizar.DataAtual   = reader.GetDateTime(reader.GetOrdinal("dataAtual"));
+                    while (reader.Read())
+                    {
+                        _atualizar.Dia = reader.GetInt16(reader.GetOrdinal("atuDiaSema"));
+                        _atualizar.Hora = Convert.ToInt16(reader.GetString(reader.GetOrdinal("atuHoraMin")).Split(':')[0]);
+                        _atualizar.Minuto = Convert.ToInt16(reader.GetString(reader.GetOrdinal("atuHoraMin")).Split(':')[1]);
+                        _atualizar.UltimaData = reader.GetDateTime(reader.GetOrdinal("atuDatAtua"));
+                        _atualizar.Versao = reader.GetInt32(reader.GetOrdinal("versao"));
+                        _atualizar.DataAtual = reader.GetDateTime(reader.GetOrdinal("dataAtual"));
+                    }
                 }
+                context.Close();
             }
-            context.Close();
+            catch (Exception e)
+            {
+                log.Error(e.Message);
+            }
+
         }
 
         /// <summary>
@@ -104,52 +114,74 @@ namespace Posto.Win.Update.Infraestrutura
         /// </summary>
         private void SetTime()
         {
-            this.AtualizaViewModel();
+            try
+            {
+                AtualizaViewModel();
 
-            var milisegundos = (this._atualizar.GetDataProximaAtualizacao.Value - this._atualizar.DataAtual).TotalMilliseconds;
+                var milisegundos = (_atualizar.GetDataProximaAtualizacao.Value - _atualizar.DataAtual).TotalMilliseconds;
 
-            this._timerProximaAtualizacao.Interval = milisegundos;
-            this._timerProximaAtualizacao.Start();
+                _timerProximaAtualizacao.Interval = milisegundos;
+                _timerProximaAtualizacao.Start();
+            }
+            catch (Exception e)
+            {
+                log.Error(e.Message);
+            }
         }
-        
-        private void Execute(object sender, ElapsedEventArgs e)
+
+        private async void Execute(object sender, ElapsedEventArgs e)
         {
-            this.ExecuteAsync();
+            await ExecuteAsync();
         }
 
         private async Task ExecuteAsync()
         {
             await Task.Run(() =>
             {
-                // INDICA AO GENEXUS MANUTENÇÃO ATIVADA E DESCONECTA TODOS OS USUÁRIOS DO BANCO.
-                this.ManutencaoAtiva();
 
-                var context  = new PostoContext(this._configuracaoModel);
-                
+                // INDICA AO GENEXUS MANUTENÇÃO ATIVADA E DESCONECTA TODOS OS USUÁRIOS DO BANCO.
+                ManutencaoAtiva();
+
+                var context = new PostoContext(_configuracaoModel);
+
                 try
                 {
                     context.BeginTransaction();
 
-                    this._atualizar.MensagemStatus = "Iniciando atualização...";
+                    _atualizar.MensagemStatus = "Iniciando atualização...";
 
-                    Process[] processes = Process.GetProcessesByName("uLeitor");
-                    foreach (Process process in processes)
+                    //-- Encerra o processo do leitor de bombas caso parâmetro seja true
+                    if (_configuracaoModel.LeitorBomba)
                     {
-                        process.Kill();
+                        Process[] processes = Process.GetProcessesByName("uLeitor");
+                        foreach (Process process in processes)
+                        {
+                            process.Kill();
+                        }
                     }
 
-                    this._timerProximaAtualizacao.Stop();
-                    this.AtualizaViewModel();
+                    //-- Encerra o processo do posto web caso parâmetro seja true
+                    if (_configuracaoModel.PostoWeb)
+                    {
+                        Process[] processes = Process.GetProcessesByName("uPostoWe");
+                        foreach (Process process in processes)
+                        {
+                            process.Kill();
+                        }
+                    }
 
-                    var ultimaVersao = this.AtualizarSql(context);
+                    _timerProximaAtualizacao.Stop();
+                    AtualizaViewModel();
 
-                    this.AtualizarExe();
-                    this.AtualizarBanco(context, ultimaVersao);
-                    
+                    var ultimaVersao = AtualizarSql(context);
+
+                    AtualizarExe();
+                    AtualizarBanco(context, ultimaVersao);
+
                     // SALVA
                     context.Commit();
 
-                    this._atualizar.MensagemStatus = "Atualizado com sucesso.";
+                    _atualizar.MensagemStatus = "Atualizado com sucesso.";
 
                 }
                 catch (Exception e)
@@ -157,24 +189,40 @@ namespace Posto.Win.Update.Infraestrutura
                     context.RollBack();
                     log.Error(e.Message);
 
-                    this._atualizar.MensagemStatus = "Problemas ao atualizar.";
+                    _atualizar.MensagemStatus = "Problemas ao atualizar.";
                 }
-                finally 
+                finally
                 {
                     context.Close();
 
-                    this._timerProximaAtualizacao.Stop();
-                    this.SetTime();
-                    this.ManutencaoDesativa(); // INDICA AO GENEXUS MANUTENÇÃO DESATIVADA
+                    _timerProximaAtualizacao.Stop();
+                    SetTime();
 
-                    //INICIA O PROCESSO DO LEITOR DE BOMBAS
-                    if (File.Exists(PathLeitor)) 
+                    //-- Indica fim de atualização
+                    ManutencaoDesativa();
+
+                    //-- Inicia processo do leitor de bombas caso parâmetro seja true
+                    if (_configuracaoModel.LeitorBomba)
                     {
-                        Process.Start(PathLeitor);
+                        //-- Se diretório existir, inicia processo
+                        if (File.Exists(PathLeitor))
+                        {
+                            Process.Start(PathLeitor);
+                        }
+                    }
+
+                    //-- Inicia processo do posto web caso parâmetro seja true
+                    if (_configuracaoModel.PostoWeb)
+                    {
+                        //-- Se diretório existir, inicia processo
+                        if (File.Exists(PathPostoWeb))
+                        {
+                            Process.Start(PathPostoWeb);
+                        }
                     }
                 }
             });
-            
+
         }
 
         /// <summary>
@@ -182,30 +230,39 @@ namespace Posto.Win.Update.Infraestrutura
         /// </summary>
         private void AtualizarExe()
         {
-            this._atualizar.MensagemStatus = "Atualizando programas...";
-
-            var arquivos = this._ftp.Download(PathExe);
-
-            using (MemoryStream mem = new MemoryStream(arquivos))
-            using (ZipArchive zipStream = new ZipArchive(mem))
+            try
             {
-                foreach (ZipArchiveEntry file in zipStream.Entries)
+                _atualizar.MensagemStatus = "Atualizando programas...";
+
+                var arquivos = _ftp.Download(PathExe);
+
+                using (MemoryStream mem = new MemoryStream(arquivos))
+                using (ZipArchive zipStream = new ZipArchive(mem))
                 {
-                    string completeFileName = Path.Combine(this._configuracaoModel.LocalDiretorio, file.FullName);
-                    if (file.Name == "")
-                    {// Assuming Empty for Directory
-                        Directory.CreateDirectory(Path.GetDirectoryName(completeFileName));
-                        continue;
+                    foreach (ZipArchiveEntry file in zipStream.Entries)
+                    {
+                        string completeFileName = Path.Combine(_configuracaoModel.LocalDiretorio, file.FullName);
+
+                        if (file.Name == "")
+                        {// Assuming Empty for Directory
+                            Directory.CreateDirectory(Path.GetDirectoryName(completeFileName));
+                            continue;
+                        }
+                        file.ExtractToFile(completeFileName, true);
                     }
-                    file.ExtractToFile(completeFileName, true);
                 }
+            }
+            catch (Exception e)
+            {
+                log.Error(e.Message);
+                _atualizar.MensagemStatus = "Problemas ao atualizar programas.";
             }
         }
 
         /// <summary>
         /// Atualiza SQLs e retorna a ultima versão
         /// </summary>
-        private int? AtualizarSql(PostoContext context) 
+        private int? AtualizarSql(PostoContext context)
         {
             var retornoSql = this._ftp.GetFileList(PathSql);
             var listaSql = retornoSql.Where(row => row.EndsWith(".sql", StringComparison.OrdinalIgnoreCase))
@@ -214,15 +271,15 @@ namespace Posto.Win.Update.Infraestrutura
                                             Versao = Convert.ToInt32(Regex.Replace(row, "[^0-9]+", "")),
                                             Arquivo = row,
                                         })
-                                        .Where(row => row.Versao > this._atualizar.Versao)
+                                        .Where(row => row.Versao > _atualizar.Versao)
                                         .OrderBy(row => row.Versao)
                                         .ToList();
 
             listaSql.ForEach(row =>
             {
-                this._atualizar.MensagemStatus = "Atualizando Rmenu " + row.Versao.ToString() + "...";
+                _atualizar.MensagemStatus = "Atualizando Rmenu " + row.Versao.ToString() + "...";
 
-                var sql = Encoding.ASCII.GetString(this._ftp.Download(PathSql + row.Arquivo));
+                var sql = Encoding.ASCII.GetString(_ftp.Download(PathSql + row.Arquivo));
                 context.Query(sql).ExecuteNonQuery();
             });
 
@@ -230,40 +287,64 @@ namespace Posto.Win.Update.Infraestrutura
                                        .Select(x => x.Versao)
                                        .FirstOrDefault();
 
-            return ultimaVersao > 0 ? ultimaVersao : this._atualizar.Versao;
+            return ultimaVersao > 0 ? ultimaVersao : _atualizar.Versao;
         }
-        
+
         /// <summary>
         /// Atualiza informações do banco
         /// </summary>
         private void AtualizarBanco(PostoContext context, int? versao)
         {
-            this._atualizar.MensagemStatus = "Atualizando versão...";
-
-            if (versao != null)
+            try
             {
-                context.Query("UPDATE oft000 SET parversao = " + versao.ToString()).ExecuteNonQuery();
-            }
+                _atualizar.MensagemStatus = "Atualizando versão...";
 
-            context.Query("UPDATE atualiz SET atuDatAtua = '" + DateTime.Now.ToString("yyyy-MM-dd hh:mm:ss") + "'").ExecuteNonQuery();
+                if (versao != null)
+                {
+                    context.Query("UPDATE oft000 SET parversao = " + versao.ToString()).ExecuteNonQuery();
+                }
+
+                context.Query("UPDATE atualiz SET atuDatAtua = '" + DateTime.Now.ToString("yyyy-MM-dd hh:mm:ss") + "'").ExecuteNonQuery();
+            }
+            catch (Exception e)
+            {
+                log.Error(e.Message);
+                _atualizar.MensagemStatus = "Problemas ao atualizar versão.";
+            }
         }
 
-        private void ManutencaoAtiva() 
+        private void ManutencaoAtiva()
         {
-            this._atualizar.MensagemStatus = "Atualizando dados...";
+            try
+            {
+                _atualizar.MensagemStatus = "Atualizando dados...";
 
-            var context = new PostoContext(this._configuracaoModel);
-            context.Query("UPDATE atualiz SET atumanuten = 'S'").ExecuteNonQuery();
-            context.Query("SELECT pg_terminate_backend(PID) FROM pg_stat_activity WHERE PID <> pg_backend_pid()").ExecuteNonQuery();
-            context.Close();
-            
+                var context = new PostoContext(_configuracaoModel);
+                context.Query("UPDATE atualiz SET atumanuten = 'S'").ExecuteNonQuery();
+                context.Query("SELECT pg_terminate_backend(PID) FROM pg_stat_activity WHERE PID <> pg_backend_pid()").ExecuteNonQuery();
+                context.Close();
+            }
+            catch (Exception e)
+            {
+                log.Error(e.Message);
+                _atualizar.MensagemStatus = "Problemas ao atualizar dados.";
+            }
+
         }
 
         private void ManutencaoDesativa()
         {
-            var context = new PostoContext(this._configuracaoModel);
-            context.Query("UPDATE atualiz SET atumanuten = 'N'").ExecuteNonQuery();
-            context.Close();
+            try
+            {
+                var context = new PostoContext(this._configuracaoModel);
+                context.Query("UPDATE atualiz SET atumanuten = 'N'").ExecuteNonQuery();
+                context.Close();
+            }
+            catch (Exception e)
+            {
+                log.Error(e.Message);
+                _atualizar.MensagemStatus = "Problemas ao finalizar manutenção.";
+            }
         }
     }
 }
